backend/requirements.txt
fastapi
tensorflow
numpy
uvicorn 
pillow